Converted weight input from integer to float

+++ weight_conversion.rb
@@ -10,7 +10,7 @@ user_input = "none"
 
 until user_input == "valid"
   
-  pounds = gets.to_i  
+  pounds = gets.to_f 
 
   if pounds <= 0 then
     puts "Does not compute. Please enter a weight greater than 0."
